fix(rpc/v05/estimate_fee): return ContractErrorV05 for reverted transactions
