+++ crates/rpc/src/v05/method.rs
@@ -1,4 +1,5 @@
 pub(crate) mod call;
+pub(crate) mod estimate_fee;
 mod get_block_with_tx_hashes;
 mod get_block_with_txs;
 mod get_transaction_receipt;
@@ -9,6 +10,7 @@ mod trace_block_transactions;
 mod trace_transaction;
 
 pub(crate) use call::call;
+pub(crate) use estimate_fee::estimate_fee;
 pub(crate) use get_block_with_tx_hashes::get_block_with_tx_hashes;
 pub(crate) use get_block_with_txs::get_block_with_txs;
 pub(crate) use get_transaction_receipt::get_transaction_receipt;
